# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для сервиса бронирований с поддержкой 2FA, уведомлений и Outbox\n\n## Возможности:\n- Регистрация и авторизация пользователей\n- Двухфакторная аутентификация (TOTP)\n- Создание и управление бронированиями\n- Система уведомлений (email, SMS, web)\n- Роли: client, staff, admin\n- Кэширование через Redis\n- Паттерн Outbox для надёжной доставки уведомлений",
        "title": "BookingGo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/UniversityProgramm/BookingGo"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/adminPanel/bookings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех бронирований в системе. Доступно только staff и admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "management"
                ],
                "summary": "Получить все бронирования",
                "responses": {
                    "200": {
                        "description": "Список всех бронирований",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/BookingGo_internal_entity.Booking"
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/me/bookings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех бронирований текущего пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Получить мои бронирования",
                "responses": {
                    "200": {
                        "description": "Список бронирований",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/BookingGo_internal_entity.Booking"
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт новое бронирование на указанный слот (1 час). Автоматически отправляет уведомление.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Создать бронирование",
                "parameters": [
                    {
                        "description": "Данные бронирования",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/BookingGo_internal_entity.CreateBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданное бронирование",
                        "schema": {
                            "$ref": "#/definitions/BookingGo_internal_entity.Booking"
                        }
                    },
                    "400": {
                        "description": "Неверный формат / время в прошлом / слот занят",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/me/bookings/cancelBooking/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Отменяет бронирование, если оно ещё не началось",
                "tags": [
                    "bookings"
                ],
                "summary": "Отменить бронирование",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Бронирование отменено"
                    },
                    "400": {
                        "description": "ID должен быть числом / бронирование уже активно",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Бронирование не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/staffPanel/bookings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех бронирований в системе. Доступно только staff и admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "management"
                ],
                "summary": "Получить все бронирования",
                "responses": {
                    "200": {
                        "description": "Список всех бронирований",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/BookingGo_internal_entity.Booking"
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/staffPanel/bookings/completeBooking/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Помечает бронирование как завершённое. Доступно только staff.",
                "tags": [
                    "management"
                ],
                "summary": "Завершить бронирование",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновлённое бронирование",
                        "schema": {
                            "$ref": "#/definitions/BookingGo_internal_entity.Booking"
                        }
                    },
                    "400": {
                        "description": "ID должен быть числом / бронирование ещё не завершено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Бронирование не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "BookingGo_internal_entity.Booking": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "problem_description": {
                    "type": "string"
                },
                "slot_end": {
                    "type": "string"
                },
                "slot_start": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/BookingGo_internal_enum.BookingStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "BookingGo_internal_entity.CreateBookingRequest": {
            "type": "object",
            "required": [
                "slot_start"
            ],
            "properties": {
                "problem_description": {
                    "type": "string"
                },
                "slot_start": {
                    "type": "string"
                }
            }
        },
...
        "BookingGo_internal_enum.BookingStatus": {
            "type": "string",
            "enum": [
                "confirmed",
                "cancelled",
                "completed"
            ],
            "x-enum-varnames": [
                "StatusConfirmed",
                "StatusCancelled",
                "StatusCompleted"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: `Bearer \u003cyour_token\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "description": "Альтернативная аутентификация через cookie",
            "type": "apiKey",
            "name": "token",
            "in": "cookie"
        }
    }
}